feitos exercicios 1 e 2
